BREAKING CHANGE! Python support 3.7+
Drop 2 and <3.5. Full support: Python >=3.7
Python 3.5-3.6 in gray zone, not tested, low effort support for a while.

Internal copy of socksipy module deleted. Install pysocks, it's in CI tests. pysocks is optional dependency.

(chore) regenerated test CA with key usage extension due to Python 3.13 SSL changes.
(chore) test dependency cryptography
