+++ .gitignore
@@ -7,6 +7,7 @@
 .cache/
 .coverage
 .hg
+.eggs/
 .installed.cfg
 .mr.developer.cfg
 .project
