+++ httplib2/error.py
