+++ httplib2/iri2uri.py
