+++ requirements-test.txt
@@ -1,16 +1,14 @@
-cryptography==38.0.3;python_version>="3.6"
+cryptography==43.0.1
 flake8==3.9.2;python_version<"3.9"
 flake8==7.3.0;python_version>="3.9"
 future==0.18.3
-mock==2.0.0;python_version<"3.3"
+pysocks==1.7.1
 pytest-cov==2.12.1
 pytest-forked==1.3.0
 pytest-randomly==1.2.3
-pytest-timeout==1.4.2;python_version<"3.7"
-pytest-timeout==2.3.1;python_version>="3.7"
+pytest-timeout==1.4.2;python_version<"3.9"
+pytest-timeout==2.3.1;python_version>="3.9"
 pytest-xdist==1.34.0
-pytest==4.6.11;python_version<="3.4"
-pytest==6.1.2;python_version>"3.4" and python_version<"3.10"
-pytest==6.2.5;python_version>="3.10" and python_version<"3.11"
-pytest==8.2.2;python_version>="3.11"
-six==1.16.0
+pytest==6.2.5;python_version<"3.9"
+pytest==7.4.4;python_version>="3.9" and python_version<"3.10"
+pytest==8.2.2;python_version>="3.10"
