+++ setup.cfg
@@ -6,7 +6,7 @@
 [coverage:run]
 
 [flake8]
-exclude = *.egg*,.env,.git,.tox,_*,build*,dist*,venv*,python2/,python3/
+exclude = *.egg*,.env,.git,.tox,_*,build*,dist*,venv*
 ignore = E261,E731,W503
 max-line-length = 121
