+++ tests/tls/ca_unused.pem
@@ -1,8 +1,8 @@
 -----BEGIN CERTIFICATE-----
-MIIDtzCCAp+gAwIBAgIUIW5VFYqUg3cVFeDgZgcCueH9P0EwDQYJKoZIhvcNAQEL
+MIIDxzCCAq+gAwIBAgIUQbWqK/j1d1znZ6HgqcFw2pUxhGwwDQYJKoZIhvcNAQEL
 BQAwazELMAkGA1UEBhMCWloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNV
 BAoMDWh0dHBsaWIyLXRlc3QxCjAIBgNVBAsMAS4xIDAeBgNVBAMMF2h0dHBsaWIy
-LXRlc3QtQ0EtdW51c2VkMB4XDTIzMDExODA2MDYwM1oXDTQzMDExMzA2MDYwM1ow
+LXRlc3QtQ0EtdW51c2VkMB4XDTI1MDcyMzE3Mjk1M1oXDTQ1MDcxODE3Mjk1M1ow
 azELMAkGA1UEBhMCWloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNVBAoM
 DWh0dHBsaWIyLXRlc3QxCjAIBgNVBAsMAS4xIDAeBgNVBAMMF2h0dHBsaWIyLXRl
 c3QtQ0EtdW51c2VkMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA1z5C
@@ -11,12 +11,13 @@ sWo94BuHBu6drgt/QxSumnvkuYbnyzPIP7cVUmUDVgg3UlPcy9w7zIQiQkit7Ozd
 YDddiXS53uSsBbeb6C8AJDiPskXDdLP+Wy7IuZODhy1MvmbOnpEUG4iC85aWoyBh
 2EBwZIFe4paj+o/oeeqPdna4FRHkMdZHbsz1WS22VfEoM5+I8yoJQJd9HLLrFGCH
 yTsbVZgUa3OBqgPOB0DHxgCbsBTKnG8Q/M+pPRBz3U/uqabKkpCCGfXUkxLFb2/A
-LQePEpHO+Ahf4YZfuwIDAQABo1MwUTAdBgNVHQ4EFgQU8KyvjPkGjGk7tq8fgWVZ
+LQePEpHO+Ahf4YZfuwIDAQABo2MwYTAdBgNVHQ4EFgQU8KyvjPkGjGk7tq8fgWVZ
 cAGGlFQwHwYDVR0jBBgwFoAU8KyvjPkGjGk7tq8fgWVZcAGGlFQwDwYDVR0TAQH/
-BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAQEAuxH5S7PkH4/nlrbZ0qSSutdvLe/O
-uCxGcyu/uQ7HkUExrLC1+9rmcJHV3F1RQ8NbJY1oLJ5jIO3EQYsw0Uh3N49n+74e
-obRrFSTJIDni2fEWYJROC7yKnFBXaU7qLcZBOnKq5cD0AlJDunTJJiD3pP5hK9E6
-P2le7TgI9sE/CIptE3Bnr+P66igsUyfDg3CF2CNGsUj30njHATcwjBJ+u2uOwHZi
-kB7v8osGtoeW0Oz54uOx9f6CuN0TbGYtaDaFaWyxnyl2lT1I+3ter9pXZUxz5NoZ
-kpW7jIXO7qSaBi1mW3ALQUoS9AA4dczJyglOuss66jsQKIz0FiYhwBz9Pg==
+BAUwAwEB/zAOBgNVHQ8BAf8EBAMCAoQwDQYJKoZIhvcNAQELBQADggEBACicFoVv
+mYhHLv4UsKHA3rMqsAyxK5zh4JdKXDu+xucw5t/1iYQg1unPLWFh0S63ueTpJ0Sf
+FSDp4BKTxaSsgpwxmBPLd9qsnnFh6liZYLtgE652dR5FmzcFnZ/c8wkGJkIYxY1K
+CuDJv1CjjOlitdnXA6iRimOJZpfO3hX8npIJiObtQtLEriJtveMGb+fk1HV8AHkP
+FruhQxtYdxZk46Lt/FMYSItiajmFh0gPSt6dDm6E45Vq29EV6sIOp/Rec1BX4JEO
+1jSM46rAZFvX6MvErbCnXY68ycaG/wFxZh0crND8vsyIG0tjWcDxqx5RZz7HlB9P
+Q2LEUoM9Wfmll1c=
 -----END CERTIFICATE-----
