+++ tests/tls/client.crt
@@ -1,22 +1,24 @@
-serial=63586944FC51DFAA087011CB861C96A4BD60AA14
-SHA1 Fingerprint=51:31:5B:9C:48:AB:D1:E6:E8:94:E6:09:E0:18:92:C7:9C:F3:01:CF
+serial=63586944FC51DFAA087011CB861C96A4BD60AA17
+SHA1 Fingerprint=33:3A:02:99:A5:58:30:41:0F:76:A8:C6:BC:B8:8D:54:BF:32:17:D3
 -----BEGIN CERTIFICATE-----
-MIIDUzCCAjsCFGNYaUT8Ud+qCHARy4YclqS9YKoUMA0GCSqGSIb3DQEBCwUAMGQx
-CzAJBgNVBAYTAlpaMQowCAYDVQQIDAEuMQowCAYDVQQHDAEuMRYwFAYDVQQKDA1o
-dHRwbGliMi10ZXN0MQowCAYDVQQLDAEuMRkwFwYDVQQDDBBodHRwbGliMi10ZXN0
-LUNBMB4XDTIzMDExODA2MDYwM1oXDTQzMDExMzA2MDYwM1owaDELMAkGA1UEBhMC
-WloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNVBAoMDWh0dHBsaWIyLXRl
-c3QxCjAIBgNVBAsMAS4xHTAbBgNVBAMMFGh0dHBsaWIyLXRlc3QtY2xpZW50MIIB
-IjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAySjxv6TnfLDZqiXA8SRPxy8c
-hPFONOxSc+uBBB8/IP1JrnUwn+lC7KEpz9TnjTlWlQS8dQSgcRIxLJX0gfQtuW9K
-xiGmmfa+Jz9jqTFsOv9evt4NGBOu6oGWhLY4elXAVVkAGVl/P6TRfu5/gJxmlORL
-V39DFeJ/qVOQiFOJGVUgUX3KWM6E8ZYDQ9HHDL0+YGHvgtrj2CAthy3LsNjGfJYY
-1S6K3K01eVxUmAryJRgtWzFF0lkZF1+ZQSE6g8Iw+LVswaOOo7seHzqZB1dDe+Uy
-az0TcyEsyCZqEBRFmRUp/cTxd4Qq0WbbOmU81X60pJ+ZQtT4dhtnFa2JGpkJUQID
-AQABMA0GCSqGSIb3DQEBCwUAA4IBAQB9knbVSHRvefkpA5DGFPV1j+jvVzg/zIeR
-/gqSZOQbbbwbQo0isR05swCFnb6yGJPml8cJMuQHzEjUUUCOOfRHkgpt5iukoEBY
-RYIm8X/C5z5X7nEIa1VhvwFotI0Eq3tpVvs8/0wzQflR4fo1qRTBXsueDBDuji7W
-8j4EwOvXrapc/wR/+1oVlTPUIQV7f5QKq/98tlTcsu0sLq2gpS1fA7VFDekKHaxb
-P7IlcCS8r8PI0ZLdlQORoxs85HJ3fcyIqILkhpw8QulfM9vVGWMBgr33L+KiIGKO
-wnjtAG1vEkfCROAQyxhY8YeJi6utziPUKl3yEBsE8EE3J9bNNNtG
+MIIDnDCCAoSgAwIBAgIUY1hpRPxR36oIcBHLhhyWpL1gqhcwDQYJKoZIhvcNAQEL
+BQAwZDELMAkGA1UEBhMCWloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNV
+BAoMDWh0dHBsaWIyLXRlc3QxCjAIBgNVBAsMAS4xGTAXBgNVBAMMEGh0dHBsaWIy
+LXRlc3QtQ0EwHhcNMjUwNzIzMTcyOTU0WhcNNDUwNzE4MTcyOTU0WjBoMQswCQYD
+VQQGEwJaWjEKMAgGA1UECAwBLjEKMAgGA1UEBwwBLjEWMBQGA1UECgwNaHR0cGxp
+YjItdGVzdDEKMAgGA1UECwwBLjEdMBsGA1UEAwwUaHR0cGxpYjItdGVzdC1jbGll
+bnQwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDJKPG/pOd8sNmqJcDx
+JE/HLxyE8U407FJz64EEHz8g/UmudTCf6ULsoSnP1OeNOVaVBLx1BKBxEjEslfSB
+9C25b0rGIaaZ9r4nP2OpMWw6/16+3g0YE67qgZaEtjh6VcBVWQAZWX8/pNF+7n+A
+nGaU5EtXf0MV4n+pU5CIU4kZVSBRfcpYzoTxlgND0ccMvT5gYe+C2uPYIC2HLcuw
+2MZ8lhjVLorcrTV5XFSYCvIlGC1bMUXSWRkXX5lBITqDwjD4tWzBo46jux4fOpkH
+V0N75TJrPRNzISzIJmoQFEWZFSn9xPF3hCrRZts6ZTzVfrSkn5lC1Ph2G2cVrYka
+mQlRAgMBAAGjQjBAMB0GA1UdDgQWBBRhqkdB90rrNVS+J2z9eoXhu396LjAfBgNV
+HSMEGDAWgBTwrK+M+QaMaTu2rx+BZVlwAYaUVDANBgkqhkiG9w0BAQsFAAOCAQEA
+hL8gXBFpGRZaDPb5d0IYtoUH2RedZKPi5mMf+VVMdEmJen5vDk0rpfJ0dO0OC6HX
+HO3B9t1fwgIUGslauN2hV96NT9qb4JlN+L5tbIMgNQG6dJ/QawMIm+38IMPltA2j
+JjbXvA4KX6Ne442K02K/sznd6WWC0IjQfQPhB1Fecij9G07Pfv3xZMjIl4gq5uOy
+FNcXfr2EHyi2X/no9umDgPTkTaSzfi0IJ2dtjhzVf1Zg/9rcvbtOTsW0P9jzwCAk
+SixyyGf2NL2W7rVEXCuLpxB569ghxPa6VIckufh6Pw+pWFMlVdIe8U/SMgEHur2h
+kCLHaf+ibjHN5J+9V0XIMA==
 -----END CERTIFICATE-----
