+++ tests/tls/client_encrypted.crt
@@ -1,22 +1,24 @@
-serial=63586944FC51DFAA087011CB861C96A4BD60AA15
-SHA1 Fingerprint=8F:0E:A4:97:BC:02:F3:48:EC:CB:92:A1:5E:A3:4B:6B:78:D2:C4:F5
+serial=63586944FC51DFAA087011CB861C96A4BD60AA18
+SHA1 Fingerprint=53:C3:89:CF:1E:A1:95:30:A7:82:52:44:8A:E7:11:BB:24:16:9B:81
 -----BEGIN CERTIFICATE-----
-MIIDVzCCAj8CFGNYaUT8Ud+qCHARy4YclqS9YKoVMA0GCSqGSIb3DQEBCwUAMGQx
-CzAJBgNVBAYTAlpaMQowCAYDVQQIDAEuMQowCAYDVQQHDAEuMRYwFAYDVQQKDA1o
-dHRwbGliMi10ZXN0MQowCAYDVQQLDAEuMRkwFwYDVQQDDBBodHRwbGliMi10ZXN0
-LUNBMB4XDTIzMDExODA2MDYwM1oXDTQzMDExMzA2MDYwM1owbDELMAkGA1UEBhMC
-WloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNVBAoMDWh0dHBsaWIyLXRl
-c3QxCjAIBgNVBAsMAS4xITAfBgNVBAMMGGh0dHBsaWIyLXRlc3QtY2xpZW50LWVu
-YzCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAMko8b+k53yw2aolwPEk
-T8cvHITxTjTsUnPrgQQfPyD9Sa51MJ/pQuyhKc/U5405VpUEvHUEoHESMSyV9IH0
-LblvSsYhppn2vic/Y6kxbDr/Xr7eDRgTruqBloS2OHpVwFVZABlZfz+k0X7uf4Cc
-ZpTkS1d/QxXif6lTkIhTiRlVIFF9yljOhPGWA0PRxwy9PmBh74La49ggLYcty7DY
-xnyWGNUuitytNXlcVJgK8iUYLVsxRdJZGRdfmUEhOoPCMPi1bMGjjqO7Hh86mQdX
-Q3vlMms9E3MhLMgmahAURZkVKf3E8XeEKtFm2zplPNV+tKSfmULU+HYbZxWtiRqZ
-CVECAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAvE6KmG1GgILd9yLQobCC8qgtN89/
-eY7isZaHGtNvbdbs1AIhwM8mQfbkBHET7lB8IAM6vRgWnek21FvOAEjZgOOKju+D
-02ylWzqqFpL2SmhS1zLOhrZxQZ7wQ1fXMGGIjNUePI9bG3eVtnKC1puj/CCDKLMX
-+O+4p2Vn51gDPN5QWNhHAKN1yqUlT9rT0zk2ahZgcYRFfwaWjlCCRjwxoP6NVvpE
-pVVZRDPuGGahTo7n69hBiBg8XcuP17brGvav0BCN8mq/oXTsa7HLLBj3vlMPTHG/
-PJt+TV9zuaMEWAdAdQDKeMLcF+nxmKi6EWE7jtiZvy9go1G6S/PVwEuPFA==
+MIIDoDCCAoigAwIBAgIUY1hpRPxR36oIcBHLhhyWpL1gqhgwDQYJKoZIhvcNAQEL
+BQAwZDELMAkGA1UEBhMCWloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNV
+BAoMDWh0dHBsaWIyLXRlc3QxCjAIBgNVBAsMAS4xGTAXBgNVBAMMEGh0dHBsaWIy
+LXRlc3QtQ0EwHhcNMjUwNzIzMTcyOTU0WhcNNDUwNzE4MTcyOTU0WjBsMQswCQYD
+VQQGEwJaWjEKMAgGA1UECAwBLjEKMAgGA1UEBwwBLjEWMBQGA1UECgwNaHR0cGxp
+YjItdGVzdDEKMAgGA1UECwwBLjEhMB8GA1UEAwwYaHR0cGxpYjItdGVzdC1jbGll
+bnQtZW5jMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAySjxv6TnfLDZ
+qiXA8SRPxy8chPFONOxSc+uBBB8/IP1JrnUwn+lC7KEpz9TnjTlWlQS8dQSgcRIx
+LJX0gfQtuW9KxiGmmfa+Jz9jqTFsOv9evt4NGBOu6oGWhLY4elXAVVkAGVl/P6TR
+fu5/gJxmlORLV39DFeJ/qVOQiFOJGVUgUX3KWM6E8ZYDQ9HHDL0+YGHvgtrj2CAt
+hy3LsNjGfJYY1S6K3K01eVxUmAryJRgtWzFF0lkZF1+ZQSE6g8Iw+LVswaOOo7se
+HzqZB1dDe+Uyaz0TcyEsyCZqEBRFmRUp/cTxd4Qq0WbbOmU81X60pJ+ZQtT4dhtn
+Fa2JGpkJUQIDAQABo0IwQDAdBgNVHQ4EFgQUYapHQfdK6zVUvids/XqF4bt/ei4w
+HwYDVR0jBBgwFoAU8KyvjPkGjGk7tq8fgWVZcAGGlFQwDQYJKoZIhvcNAQELBQAD
+ggEBAGuU7rE2GDPdGbdUWGBMmy1eBFmLlkcBT/olLMbdNnfnyAslRmPDblhysIkW
+PZTXZsM39S8jhI61fWreR5T3Fevw9Xmt/FA+jW6OJWy8DdIwkDU6+u7vlAAviuBj
+H54sHtuKxlNEX52PVljI20WNNCoW0PfhASVjMWsyw+UBWFsN+rJw2KZXnjq/ymF6
+0/3IE1A0wOFl6qehi9p/vvQFiEyxaDhXKfmDshQ5I4ILPdm9jf7JLLT7Ndh6S4Kl
+NNUY0hM1lZJANM8qmowI0XNAX/4YnxDk4OQoYgJeNteQNClxxlPNJhFZjgkNotu1
++cWeBhGhoQLJ77sDvfAZQAI2l5c=
 -----END CERTIFICATE-----
