+++ tests/tls/server.pem
@@ -1,24 +1,26 @@
-serial=63586944FC51DFAA087011CB861C96A4BD60AA16
-SHA1 Fingerprint=AD:96:2D:BE:56:E2:58:76:85:FF:34:44:0D:D5:0E:08:98:CB:EA:84
+serial=63586944FC51DFAA087011CB861C96A4BD60AA19
+SHA1 Fingerprint=67:8B:EB:99:4B:31:DF:86:38:E0:2F:1D:9A:6A:94:2C:AF:F2:F6:F7
 -----BEGIN CERTIFICATE-----
-MIIDSDCCAjACFGNYaUT8Ud+qCHARy4YclqS9YKoWMA0GCSqGSIb3DQEBCwUAMGQx
-CzAJBgNVBAYTAlpaMQowCAYDVQQIDAEuMQowCAYDVQQHDAEuMRYwFAYDVQQKDA1o
-dHRwbGliMi10ZXN0MQowCAYDVQQLDAEuMRkwFwYDVQQDDBBodHRwbGliMi10ZXN0
-LUNBMB4XDTIzMDExODA2MDYwM1oXDTQzMDExMzA2MDYwM1owXTELMAkGA1UEBhMC
-WloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNVBAoMDWh0dHBsaWIyLXRl
-c3QxCjAIBgNVBAsMAS4xEjAQBgNVBAMMCWxvY2FsaG9zdDCCASIwDQYJKoZIhvcN
-AQEBBQADggEPADCCAQoCggEBAMP5DQnuT8bdUct4d8WCHmBfTiGqEmZGxrGet7fL
-tAbxYHTJmXB+Qv8rR/naMpzqN5Bs1p6/3dByl0C0r4LnyoI/wPVy4NuYPGOKUHKI
-9ET4BJvS63/me6jlmBgrZHpvEhA/gUGmEgiOa9vLzkKrV0NB22h01Vn6TNKosia3
-qFL8Q9t9eSU4bz0bhwl3KJZYWEEumiUBxCeIT+h7i9lqsZG5ch3egpPOD7fAasT7
-lxRZjeVfVMyMkgyo6TwKAhLNCMsQFCW5UIrvob0/p2xYkx+UDBTdBkLOaaKC1aOa
-++nXk5PYHQiWaQK6z37ENS/IEq86lbnewCtHeuGPBWaRe5kCAwEAATANBgkqhkiG
-9w0BAQsFAAOCAQEAQF4e+y9Jq/WUredh09af3T4tIRQoAXDT08TE0qBTlcJQxA63
-5tIkK+mY8S80E+/2FWcYquJxQGOdC2ymGWolVgj5hL0cmgq468UAi9M4Op/bXHjY
-MMk4cg79folhDtGAXBIDWR9lZ+VUVKBXYUJAsfAm6fKgKTDzSkZu1w30viBjPA8v
-Cw5XZpesGcz/qvrRAOJs+S5IKLAJBnRHIorT9koOreJclGW0vxtRVy1OKuxK3mIv
-Z3J4OhEyX59dbWyafCbfP80TUInieV+6Ij+3SDC7aA7uaJjV2gjet2yRjL5vFvYD
-tYQl1L6G3UNQX6UuUzE1tx80Mxg33RsHVILLoA==
+MIIDkTCCAnmgAwIBAgIUY1hpRPxR36oIcBHLhhyWpL1gqhkwDQYJKoZIhvcNAQEL
+BQAwZDELMAkGA1UEBhMCWloxCjAIBgNVBAgMAS4xCjAIBgNVBAcMAS4xFjAUBgNV
+BAoMDWh0dHBsaWIyLXRlc3QxCjAIBgNVBAsMAS4xGTAXBgNVBAMMEGh0dHBsaWIy
+LXRlc3QtQ0EwHhcNMjUwNzIzMTcyOTU0WhcNNDUwNzE4MTcyOTU0WjBdMQswCQYD
+VQQGEwJaWjEKMAgGA1UECAwBLjEKMAgGA1UEBwwBLjEWMBQGA1UECgwNaHR0cGxp
+YjItdGVzdDEKMAgGA1UECwwBLjESMBAGA1UEAwwJbG9jYWxob3N0MIIBIjANBgkq
+hkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAw/kNCe5Pxt1Ry3h3xYIeYF9OIaoSZkbG
+sZ63t8u0BvFgdMmZcH5C/ytH+doynOo3kGzWnr/d0HKXQLSvgufKgj/A9XLg25g8
+Y4pQcoj0RPgEm9Lrf+Z7qOWYGCtkem8SED+BQaYSCI5r28vOQqtXQ0HbaHTVWfpM
+0qiyJreoUvxD2315JThvPRuHCXcollhYQS6aJQHEJ4hP6HuL2WqxkblyHd6Ck84P
+t8BqxPuXFFmN5V9UzIySDKjpPAoCEs0IyxAUJblQiu+hvT+nbFiTH5QMFN0GQs5p
+ooLVo5r76deTk9gdCJZpArrPfsQ1L8gSrzqVud7AK0d64Y8FZpF7mQIDAQABo0Iw
+QDAdBgNVHQ4EFgQUe+wkwVu8XFjzYSv6t297d8hGK58wHwYDVR0jBBgwFoAU8Kyv
+jPkGjGk7tq8fgWVZcAGGlFQwDQYJKoZIhvcNAQELBQADggEBAKwIUmj/kOhSpAvI
+p3O+JDYG8VhuVj3gCU1mjMi5lmMMNZ+Zj8/snwOnfQlbQ/fLTyO8gK7sZgTRCcPq
+k1jcSjhSuVbefq2h+k7xrO3rFCfuIeXiM1QM6LpD5UJnvuUePNT6NbOsHl6hG93i
+YM6m+/OtUU2NNMMYbOrFNIo3cW81JVw28O1N6ez4E3F676aiPuKQS8Ip5Yi2YZIX
+64pqhXFfT2MUg0TaQv9Xw2/jNKgdjaWNFHPcUryR56QXFTD0jm9lHQ3dXG8eGLAQ
+I2vkT0D6Yi+43syjJg/MWta+UXY9JAljzgtlNHYSGym50xqkQ+glAVUzsiuTnURM
+a3ixs18=
 -----END CERTIFICATE-----
 -----BEGIN PRIVATE KEY-----
 MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDD+Q0J7k/G3VHL
